ml-service/requirements.txt
# FastAPI & server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ML & Data Science - State of the Art
scikit-learn==1.6.1
xgboost==2.1.3
catboost==1.2.7
lightgbm==4.5.0
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2

# Clustering - SOTA
hdbscan==0.8.39

# Data augmentation - SMOTE for class imbalance
imbalanced-learn==0.14.1

# Time series forecasting - LightGBM with Islamic calendar covariates
hijri-converter==2.3.2

# LLM Integration
google-genai>=1.0.0

# Database
mysql-connector-python==9.1.0

# Configuration
python-dotenv==1.0.1

# HTTP client
httpx==0.28.1

# Visualization
matplotlib==3.9.4
seaborn==0.13.2
